+++ include/pbvp/recreate_test.hpp
@@ -3,5 +3,6 @@
 namespace pbvp::diagnostics {
 
 bool ScheduleEngineRecreateTest() noexcept;
+void ObserveEngineRecreateTest() noexcept;
 
 } // namespace pbvp::diagnostics